+++ smac/multi_objective/parego.py
@@ -37,8 +37,8 @@ class ParEGO(AbstractMultiObjectiveAlgorithm):
         self._rng = np.random.RandomState(seed)
 
         self._rho = rho
-        self._theta = self._rng.rand(self._n_objectives)
-        self.update_on_iteration_start()
+        # Will be set on starting an SMBO iteration
+        self._theta: np.ndarray | None = None
 
     @property
     def meta(self) -> dict[str, Any]:  # noqa: D102
@@ -61,5 +61,8 @@ class ParEGO(AbstractMultiObjectiveAlgorithm):
 
     def __call__(self, values: list[float]) -> float:  # noqa: D102
         # Weight the values
+        if self._theta is None:
+            raise ValueError("Iteration not yet initalized; Call `update_on_iteration_start()` first")
+
         theta_f = self._theta * values
         return float(np.max(theta_f, axis=0) + self._rho * np.sum(theta_f, axis=0))