+++ smac/__init__.py
@@ -19,7 +19,7 @@ copyright = f"""
     Copyright {datetime.date.today().strftime('%Y')}, Marius Lindauer, Katharina Eggensperger,
     Matthias Feurer, André Biedenkapp, Difan Deng, Carolin Benjamins, Tim Ruhkopf, René Sass
     and Frank Hutter"""
-version = "2.0.0a2"
+version = "2.0.0a3"
 
 
 try:
@@ -28,7 +28,9 @@ try:
     setup_logging(0)
 
     from smac.callback import Callback
-    from smac.facade import (  # HyperbandFacade,; MultiFidelityFacade,
+    from smac.facade import (
+        HyperbandFacade,
+        MultiFidelityFacade,
         AlgorithmConfigurationFacade,
         BlackBoxFacade,
         HyperparameterOptimizationFacade,