+++ smac/facade/__init__.py
@@ -2,12 +2,12 @@ from smac.facade.abstract_facade import AbstractFacade
 from smac.facade.algorithm_configuration_facade import AlgorithmConfigurationFacade
 from smac.facade.blackbox_facade import BlackBoxFacade
 
-# from smac.facade.hyperband_facade import HyperbandFacade
+from smac.facade.hyperband_facade import HyperbandFacade
 from smac.facade.hyperparameter_optimization_facade import (
     HyperparameterOptimizationFacade,
 )
 
-# from smac.facade.multi_fidelity_facade import MultiFidelityFacade
+from smac.facade.multi_fidelity_facade import MultiFidelityFacade
 from smac.facade.random_facade import RandomFacade
 
 __all__ = [
@@ -15,7 +15,7 @@ __all__ = [
     "AlgorithmConfigurationFacade",
     "BlackBoxFacade",
     "HyperparameterOptimizationFacade",
-    # "MultiFidelityFacade",
-    # "HyperbandFacade",
+    "MultiFidelityFacade",
+    "HyperbandFacade",
     "RandomFacade",
 ]